+++ README.md
@@ -16,4 +16,4 @@ Split the given `App` markup into component following the rules:
 - Install Prettier Extention and use this [VSCode settings](https://mate-academy.github.io/fe-program/tools/vscode/settings.json) to enable format on save.
 - Implement a solution following the [React task guideline](https://github.com/mate-academy/react_task-guideline#react-tasks-guideline).
 - Open one more terminal and run tests with `npm test` to ensure your solution is correct.
-- Replace `<your_account>` with your Github username in the [DEMO LINK](https://<your_account>.github.io/react_static-list-of-todos-js/) and add it to the PR description.
+- Replace `<your_account>` with your Github username in the [DEMO LINK](https://reznik-denis.github.io/react_static-list-of-todos-js/) and add it to the PR description.
