+++ src/components/TodoInfo/TodoInfo.jsx
@@ -1,2 +1,12 @@
-// Add the required props
-export const TodoInfo = () => <>TodoInfo markup</>;
+import './todoInfo.scss';
+import { UserInfo } from '../UserInfo';
+
+export const TodoInfo = ({ todo }) => (
+  <article
+    className={todo.completed ? 'TodoInfo TodoInfo--completed' : 'TodoInfo'}
+  >
+    <h2 className="TodoInfo__title">{todo.title}</h2>
+
+    {todo.user && <UserInfo user={todo.user} />}
+  </article>
+);
