+++ src/components/UserInfo/UserInfo.jsx
@@ -1,2 +1,7 @@
-// Add the required props
-export const UserInfo = () => <>UserInfo markup</>;
+import './userInfo.scss';
+
+export const UserInfo = ({ user }) => (
+  <a className="UserInfo" href={`mailto:${user.email}`}>
+    {user.name}
+  </a>
+);
