+++ src/components/UserInfo/userInfo.scss
@@ -1,0 +1,3 @@
+.UserInfo {
+  font-style: italic;
+}
